Move Events up by a namespace
