+++ src/Event/Listener/CiInitialTestsConsoleLoggerSubscriber.php
@@ -35,8 +35,8 @@ declare(strict_types=1);
 
 namespace Infection\Event\Listener;
 
-use Infection\Event\Event\InitialTestSuiteStarted;
 use Infection\Event\EventDispatcher\EventSubscriberInterface;
+use Infection\Event\InitialTestSuiteStarted;
 use Infection\TestFramework\TestFrameworkAdapter;
 use InvalidArgumentException;
 use Symfony\Component\Console\Output\OutputInterface;
