+++ src/Event/MutantCreated.php
@@ -33,11 +33,11 @@
 
 declare(strict_types=1);
 
-namespace Infection\Event\Event;
+namespace Infection\Event;
 
 /**
  * @internal
  */
-final class MutationGeneratingFinished
+final class MutantCreated
 {
 }
