+++ src/Event/MutationTestingFinished.php
@@ -33,19 +33,11 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests\Event\Event;
+namespace Infection\Event;
 
-use Infection\Event\Event\InitialTestCaseCompleted;
-use PHPUnit\Framework\TestCase;
-
-final class InitialTestCaseCompletedTest extends TestCase
+/**
+ * @internal
+ */
+final class MutationTestingFinished
 {
-    /**
-     * This class is only used to fire events, and the only functionality it needs is being instantiated
-     */
-    public function test_it_can_be_initialzed(): void
-    {
-        $class = new InitialTestCaseCompleted();
-        $this->assertInstanceOf(InitialTestCaseCompleted::class, $class);
-    }
 }
