+++ tests/phpunit/Event/ApplicationExecutionFinishedTest.php
@@ -33,19 +33,19 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests\Event\Event;
+namespace Infection\Tests\Event;
 
-use Infection\Event\Event\MutableFileProcessed;
+use Infection\Event\ApplicationExecutionFinished;
 use PHPUnit\Framework\TestCase;
 
-final class MutableFileProcessedTest extends TestCase
+final class ApplicationExecutionFinishedTest extends TestCase
 {
     /**
      * This class is only used to fire events, and the only functionality it needs is being instantiated
      */
     public function test_it_can_be_initialzed(): void
     {
-        $class = new MutableFileProcessed();
-        $this->assertInstanceOf(MutableFileProcessed::class, $class);
+        $class = new ApplicationExecutionFinished();
+        $this->assertInstanceOf(ApplicationExecutionFinished::class, $class);
     }
 }
