+++ tests/phpunit/Event/Listener/CiInitialTestsConsoleLoggerSubscriberTest.php
@@ -35,8 +35,8 @@ declare(strict_types=1);
 
 namespace Infection\Tests\Event\Listener;
 
-use Infection\Event\Event\InitialTestSuiteStarted;
 use Infection\Event\EventDispatcher\EventDispatcher;
+use Infection\Event\InitialTestSuiteStarted;
 use Infection\Event\Listener\CiInitialTestsConsoleLoggerSubscriber;
 use Infection\TestFramework\AbstractTestFrameworkAdapter;
 use PHPUnit\Framework\MockObject\MockObject;
