+++ tests/phpunit/Event/Listener/MutationTestingConsoleLoggerSubscriberTest.php
@@ -37,11 +37,11 @@ namespace Infection\Tests\Event\Listener;
 
 use Infection\Console\OutputFormatter\OutputFormatter;
 use Infection\Differ\DiffColorizer;
-use Infection\Event\Event\MutantProcessFinished;
-use Infection\Event\Event\MutationTestingFinished;
-use Infection\Event\Event\MutationTestingStarted;
 use Infection\Event\EventDispatcher\EventDispatcher;
 use Infection\Event\Listener\MutationTestingConsoleLoggerSubscriber;
+use Infection\Event\MutantProcessFinished;
+use Infection\Event\MutationTestingFinished;
+use Infection\Event\MutationTestingStarted;
 use Infection\Mutant\MetricsCalculator;
 use Infection\Process\MutantProcess;
 use PHPUnit\Framework\MockObject\MockObject;
