+++ tests/phpunit/Event/MutantsCreatingFinishedTest.php
@@ -33,11 +33,19 @@
 
 declare(strict_types=1);
 
-namespace Infection\Event\Event;
+namespace Infection\Tests\Event;
 
-/**
- * @internal
- */
-final class InitialTestCaseCompleted
+use Infection\Event\MutantsCreatingFinished;
+use PHPUnit\Framework\TestCase;
+
+final class MutantsCreatingFinishedTest extends TestCase
 {
+    /**
+     * This class is only used to fire events, and the only functionality it needs is being instantiated
+     */
+    public function test_it_can_be_initialzed(): void
+    {
+        $class = new MutantsCreatingFinished();
+        $this->assertInstanceOf(MutantsCreatingFinished::class, $class);
+    }
 }
