+++ tests/phpunit/Event/MutationGeneratingStartedTest.php
@@ -33,11 +33,18 @@
 
 declare(strict_types=1);
 
-namespace Infection\Event\Event;
+namespace Infection\Tests\Event;
 
-/**
- * @internal
- */
-final class ApplicationExecutionFinished
+use Infection\Event\MutationGeneratingStarted;
+use PHPUnit\Framework\TestCase;
+
+final class MutationGeneratingStartedTest extends TestCase
 {
+    public function test_it_passes_along_its_mutable_file_count_without_changing_it(): void
+    {
+        $count = 5;
+        $event = new MutationGeneratingStarted($count);
+
+        $this->assertSame($count, $event->getMutableFilesCount());
+    }
 }
